저장할 db의 엔티티 클래스와 저장을 수행할 레퍼지토리 클래스 생성 (+ 에이전트 서버의 필드명 db에 맞게 수정)

+++ franchise/backend/src/main/java/com/franchise/backend/briefing/dto/FocusPointCheckedDto.java
@@ -10,7 +10,7 @@ import lombok.*;
 public class FocusPointCheckedDto {
 
     @JsonProperty("check_id")
-    private Integer checkId;
+    private Long checkId;
 
     private Boolean check;
